requirements.txt
Django==1.8.5
djangorestframework==3.2.4
pydruid==0.2.2
django-datetime-widget==0.9.3
django-rest-swagger==0.3.4
pandas==0.16.2
python-dateutil==2.4.2
pytz==2015.6
PyYAML==3.11
requests==2.8.0
rest-pandas==0.3.2
six==1.10.0
wheel==0.24.0
